Safari-611.1.10/LayoutTests/editing/selection/4895428-4-expected.txt

This tests to see if an editable selection containing a focusable node is draggable by content inside that focusable node. To achieve this we delay focus of such nodes until mouse up. Below, the table should be inside the red bordered div.
| "\n"
| <div>
|   contenteditable="true"
|   id="div"
|   style="border: 1px solid blue;"
|   "\n"
|   "\n"
| "\n"
| <div>
|   contenteditable="true"
|   id="destination"
|   style="border: 1px solid red;"
|   <#selection-anchor>
|   <table>
|     border="1"
|     <tbody>
|       <tr>
|         <td>
|           contenteditable="true"
|           id="cell"
|           "editable<#selection-focus>"
| "\n"
| <ul>
|   id="console"
| "\n"
